api/requirements.txt
celery==5.0.1
Django==3.1.2
djangorestframework==3.12.1
flake8==3.8.4
gunicorn==20.0.4
mutagen==1.45.1
redis==3.5.3
psycopg2-binary==2.8.6
